+++ src/authenticator.rs
@@ -3,9 +3,6 @@
 use crate::ctap1;
 use crate::ctap2;
 
-pub use ctap1::Authenticator as Ctap1Authenticator;
-pub use ctap2::Authenticator as Ctap2Authenticator;
-
 #[derive(Clone, Debug, PartialEq)]
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
 // clippy says (2022-02-26): large size difference