+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove re-exports for `heapless` and `heapless-bytes` as the relevant types are already re-exported.
 - `ctap2::get_info`: Use subcommand enum for `authenticator_config_commands`.
 - `webauthn`: Make `KnownPublicKeyCredentialParameters` an enum.
+- `authenticator`: Move `Authenticator`, `Request` and `Response` to the crate root.
 
 ## [0.6.0-rc.4] 2026-06-01